Rename embeddable action storage
